Crate polkadot_performance_test

A Polkadot performance tests utilities.

Enums

PerfCheckError

Constants

ERASURE_CODING_N_VALIDATORS

Value used for reference benchmark of erasure-coding.

ERASURE_CODING_TIME_LIMIT

PVF_PREPARE_TIME_LIMIT

VALIDATION_CODE_BOMB_LIMIT

The bomb limit for decompressing code blobs.

WASM_BINARY

Functions

measure_erasure_coding

Measure the time it takes to break arbitrary data into chunks and reconstruct it back.

measure_pvf_prepare

Measures the time it takes to compile arbitrary wasm code.